requirements.txt
python-dotenv
openai
ipykernel
ipywidgets
gradio
